refactor: merge demo  generatelogic
